+++ src/features/atlas/countries/components/countriesPanel/CountriesPanel.tsx
@@ -2,8 +2,8 @@ import { useMemo } from "react";
 import { useTranslation } from "react-i18next";
 import { ActionButton, Panel, Separator } from "@components";
 import { ICONS } from "@constants/icons";
+import { useCountryFilters } from "@contexts/CountryFiltersContext";
 import { useCountryLists } from "@contexts/CountryListsContext";
-import { useTimeline } from "@contexts/TimelineContext";
 import { useTrips } from "@contexts/TripsContext";
 import { useUI } from "@contexts/UIContext";
 import {
@@ -16,7 +16,6 @@ import { useSort } from "@hooks";
 import { CountriesSearchSortBar } from "./CountriesSearchSortBar";
 import { CountryListView } from "./CountryListView";
 import { CountryFiltersPanel } from "../countryFilters/CountryFiltersPanel";
-import { useCountryFilters } from "../../hooks/useCountryFilters";
 
 interface CountriesPanelProps {
   selectedIsoCode: string | null;
@@ -45,9 +44,9 @@ export function CountriesPanel({
     showFilters,
     toggleFilters,
   } = useUI();
-  const { showVisitedOnly, setShowVisitedOnly } = useTimeline();
 
   // Filter state
+  const { resetFilters, ...filterProps } = useCountryFilters();
   const {
     countryLists,
     selectedListId,
@@ -55,8 +54,6 @@ export function CountriesPanel({
     openAddModal,
     openEditModal,
   } = useCountryLists();
-
-  const { resetFilters, ...filterProps } = useCountryFilters();
 
   // Sort state
   const {
@@ -131,20 +128,8 @@ export function CountriesPanel({
       >
         <div className="flex flex-col h-full">
           <CountriesSearchSortBar
-            search={filterProps.search}
-            setSearch={filterProps.setSearch}
             sortBy={sortBy}
             setSortBy={(v) => setSortBy(v as typeof sortBy)}
-            sovereignOnly={filterProps.sovereignOnly}
-            setSovereignOnly={filterProps.setSovereignOnly}
-            visitedOnly={showVisitedOnly}
-            setVisitedOnly={setShowVisitedOnly}
-            wantToVisitOnly={filterProps.wantToVisitOnly}
-            setWantToVisitOnly={filterProps.setWantToVisitOnly}
-            allCount={filterProps.allCount}
-            sovereignCount={filterProps.sovereignCount}
-            visitedCount={filterProps.visitedCount}
-            wantToVisitCount={filterProps.wantToVisitCount}
             countryLists={dynamicCountryLists}
             selectedListId={selectedListId}
             setSelectedListId={setSelectedListId}
@@ -167,10 +152,7 @@ export function CountriesPanel({
         <CountryFiltersPanel
           show={showFilters && !selectedCountry}
           onHide={toggleFilters}
-          visitedOnly={showVisitedOnly}
-          showVisitedOnly={showVisitedOnly}
           resetFilters={handleResetFilters}
-          {...filterProps}
         />
       )}
     </div>